pyproject.toml
[project]
name = "tombstone-cleanup"
version = "0.1.0"
description = "Kafka-based distributed processor"
readme = "README.md"
authors = [{ name = "Resonate HQ", email = "contact@resonatehq.io" }]
requires-python = ">=3.11"  # 3.13 is too new for many packages including uv; downgrade for stability

dependencies = [
    "click>=8.1.8",
    "kafka-python>=2.0.4",
    "resonate-sdk>=0.4.12",
]

[project.scripts]
record-producer = "record_producer.produce:main"
record-processor = "record_processor.processor:main"
setup-topics = "setup.setup_topics:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
# Optional: You can add tool-specific config here

[dependency-groups]
dev = ["pyright[nodejs]>=1.1.394", "ruff>=0.9.7"]

[tool.hatch.build.targets.wheel]
packages = ["record_producer", "record_consumer", "setup"]